backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1

# Google Gemini (new package)
google-genai==1.0.0

# Firestore
google-cloud-firestore==2.16.0
google-auth==2.28.1

# ML/Scikit-learn
scikit-learn==1.4.2
numpy==1.26.4

# Scheduler
apscheduler==3.10.4
